fix: onDevCompileDone hook async not work
