+++ packages/compat/webpack/src/core/build.ts
@@ -4,7 +4,7 @@ import {
   logger,
   getNodeEnv,
   setNodeEnv,
-  isMultiCompiler,
+  onCompileDone,
   type Stats,
   type Rspack,
   type MultiStats,
@@ -52,40 +52,7 @@ export const build = async (
     await p;
   };
 
-  // MultiCompiler does not supports `done.tapPromise`
-  if (isMultiCompiler(compiler)) {
-    const { compilers } = compiler;
-    const compilerStats: Stats[] = [];
-    let doneCompilers = 0;
-
-    for (let index = 0; index < compilers.length; index++) {
-      const compiler = compilers[index];
-      const compilerIndex = index;
-      let compilerDone = false;
-
-      compiler.hooks.done.tapPromise('rsbuild:done', async (stats) => {
-        if (!compilerDone) {
-          compilerDone = true;
-          doneCompilers++;
-        }
-
-        compilerStats[compilerIndex] = stats;
-
-        if (doneCompilers === compilers.length) {
-          await onDone(new WebpackMultiStats(compilerStats));
-        }
-      });
-
-      compiler.hooks.invalid.tap('rsbuild:done', () => {
-        if (compilerDone) {
-          compilerDone = false;
-          doneCompilers--;
-        }
-      });
-    }
-  } else {
-    compiler.hooks.done.tapPromise('rsbuild:done', onDone);
-  }
+  onCompileDone(compiler, onDone, WebpackMultiStats);
 
   if (watch) {
     compiler.watch({}, (err) => {
