+++ packages/shared/src/utils.ts
@@ -1,7 +1,9 @@
 import path from 'node:path';
 import type { Compiler, MultiCompiler } from '@rspack/core';
 import type {
+  Stats,
   NodeEnv,
+  MultiStats,
   CacheGroups,
   CompilerTapFn,
   RsbuildTarget,
@@ -358,4 +360,46 @@ export const isMultiCompiler = <
   compiler: C | M,
 ): compiler is M => {
   return compiler.constructor.name === 'MultiCompiler';
+};
+
+export const onCompileDone = (
+  compiler: Compiler | MultiCompiler,
+  onDone: (stats: Stats | MultiStats) => Promise<void>,
+  MultiStatsCtor: new (stats: Stats[]) => MultiStats,
+) => {
+  // The MultiCompiler of Rspack does not supports `done.tapPromise`,
+  // so we need to use the `done` hook of `MultiCompiler.compilers` to implement it.
+  if (isMultiCompiler(compiler)) {
+    const { compilers } = compiler;
+    const compilerStats: Stats[] = [];
+    let doneCompilers = 0;
+
+    for (let index = 0; index < compilers.length; index++) {
+      const compiler = compilers[index];
+      const compilerIndex = index;
+      let compilerDone = false;
+
+      compiler.hooks.done.tapPromise('rsbuild:done', async (stats) => {
+        if (!compilerDone) {
+          compilerDone = true;
+          doneCompilers++;
+        }
+
+        compilerStats[compilerIndex] = stats;
+
+        if (doneCompilers === compilers.length) {
+          await onDone(new MultiStatsCtor(compilerStats));
+        }
+      });
+
+      compiler.hooks.invalid.tap('rsbuild:done', () => {
+        if (compilerDone) {
+          compilerDone = false;
+          doneCompilers--;
+        }
+      });
+    }
+  } else {
+    compiler.hooks.done.tapPromise('rsbuild:done', onDone);
+  }
 };
